+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fix breadcrumb, only for main spa booking category
+
 ### 1.3.5: July 15th, 2021
 * Icons: consider theme and fontawesome